requirements.txt
config==0.4.2
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==11.1.0
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==2.8.0
gunicorn==19.9.0
Pillow==6.1.0
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.1
whitenoise==4.1.3
